PVS and build warnings fixes

+++ src/xrCore/SubAlloc.hpp
@@ -52,7 +52,7 @@ inline void SplitBlock(void* pv,UINT OldIndx,UINT NewIndx)
 }
 DWORD _STDCALL GetUsedMemory()
 {
-    DWORD i, RetVal=SubAllocatorSize-(HiUnit-LoUnit)-(UnitsStart-pText);
+    DWORD i, RetVal= DWORD(SubAllocatorSize-(HiUnit-LoUnit)-(UnitsStart-pText));
     for (i=0;i < N_INDEXES;i++)
             RetVal -= UNIT_SIZE*Indx2Units[i]*BList[i].Stamp;
     return RetVal;
